Refact UpdateTimerCallback. Move variables, and common methods in TelegramBotClient to common class part (TelegramBotClientCommon)
